third_party/WebKit/LayoutTests/animations/animationworklet/animator-registration-expected.txt

CONSOLE MESSAGE: line 5: [object DOMException]
CONSOLE MESSAGE: line 4: TypeError: Failed to execute 'registerAnimator' on 'AnimationWorkletGlobalScope': The callback provided as parameter 2 is not a function.
CONSOLE MESSAGE: line 4: TypeError: Failed to execute 'registerAnimator' on 'AnimationWorkletGlobalScope': The empty string is not a valid name.
